moved api calls to App.jsx
