+++ frontend/src/components/Resources/Resource.module.css
@@ -75,41 +75,3 @@
     font-size: 16px;
     font-variant: small-caps;
 }
-
-/* Spinner while waiting to load */
-.load_more {
-    display: flex;
-    justify-content: center;
-    margin-bottom: 30px;
-}
-.loading {
-    display: flex;
-    justify-content: center;
-    margin: 20px;
-}
-
-.spinner {
-    border: 4px solid #0F5FA6;
-    border-top: 4px solid white;
-    border-radius: 50%;
-    width: 50px;
-    height: 50px;
-    flex-shrink: 0;
-    animation: spin 1s linear infinite;
-    margin: 0 30px;
-}
-
-@keyframes spin {
-    0% { transform: rotate(0deg); }
-    100% { transform: rotate(360deg); }
-}
-
-/* Failed to load resources */
-.retry {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    align-items: center;
-    margin: 40px;
-    height: 200px;
-}
